feat(pi): extract the shared session builder (Phase 0) + unify the assembly front half
Extract the definition-aware AgentSessionRuntime builder from chat.ts
into session-builder.ts; chat becomes one consumer. The TUI-only ~/.pi
auth divergence is eliminated in place: the builder injects fastagent's
credential store via a ModelRuntime (createPiModelRuntime, the
ModelRuntime-shaped sibling of createPiModels), so pi's /login writes to
the workspace auth file and --auth-path/FASTAGENT_AUTH_PATH now apply to
chat.

Unify the workspace assembly front half (resolveWorkspaceAssembly):
config → model spec → agentDir → tool surface → state root → auth path
is computed in ONE place, consumed by both the serving opener and the
session builder — the two pi consumption shapes can no longer drift on
what they assemble. Fidelity fix surfaced by the unification: the
resident session now honors config.thinkingLevel like serving does.
