+++ src/cli/commands/chat.ts
@@ -4,7 +4,7 @@ import { loadDotEnv } from "../../env.ts";
 import { installProxyFetch } from "../../proxy.ts";
 import { failStartup } from "../fail.ts";
 
-export async function runChat(dirArg: string, opts: { model?: string }): Promise<void> {
+export async function runChat(dirArg: string, opts: { model?: string; authPath?: string }): Promise<void> {
   const dir = resolve(dirArg);
   loadDotEnv(dir);
   installProxyFetch(); // model calls (and the login dialog) must go through the proxy too
@@ -13,5 +13,5 @@ export async function runChat(dirArg: string, opts: { model?: string }): Promise
   process.chdir(dir);
   // Lazy-import: chat pulls pi's interactive TUI module graph; headless start/dev never need it.
   const { runPiChat } = await import("../../engines/pi/chat.ts");
-  await runPiChat(dir, { model: opts.model }).catch(failStartup);
+  await runPiChat(dir, { model: opts.model, authPath: opts.authPath }).catch(failStartup);
 }
