+++ test/session-builder.test.ts
@@ -4,13 +4,13 @@ import { tmpdir } from "node:os";
 import { join } from "node:path";
 import { SessionManager } from "@earendil-works/pi-coding-agent";
 import { describe, expect, it, vi } from "vitest";
-import { buildChatRuntime } from "../src/engines/pi/chat.ts";
+import { buildWorkspaceSessionRuntime } from "../src/engines/pi/session-builder.ts";
 
 // `chat` must run the SAME agent dev/start serve, presented in pi's TUI — NOT pi's vanilla
-// discovery. buildChatRuntime is split out precisely so that injection is inspectable without a
+// discovery. buildWorkspaceSessionRuntime is split out precisely so that injection is inspectable without a
 // TTY. In-memory sessions keep the test from writing to the machine's pi session store. A raw
 // AgentTool via `config.tools` lets the custom-tool path be tested without installing the package.
-describe("chat: buildChatRuntime injects fastagent's assembled agent into pi's session", () => {
+describe("session builder: buildWorkspaceSessionRuntime injects fastagent's assembled agent into pi's session", () => {
   it("emulates deferral: deferred tools start inactive, search_tools mounts and activates via pi's session", async () => {
     const dir = await mkdtemp(join(tmpdir(), "fa-chat-defer-"));
     try {
@@ -27,7 +27,7 @@ describe("chat: buildChatRuntime injects fastagent's assembled agent into pi's s
            }],
          };\n`,
       );
-      const rt = await buildChatRuntime(dir, {}, SessionManager.inMemory());
+      const rt = await buildWorkspaceSessionRuntime(dir, {}, SessionManager.inMemory());
       try {
         const session = rt.session;
         // Initial active set mirrors serving: deferred tool registered but NOT active; loader active.
@@ -123,7 +123,7 @@ describe("chat: buildChatRuntime injects fastagent's assembled agent into pi's s
       const errorSpy = vi.spyOn(console, "error").mockImplementation((...args: unknown[]) => {
         warnings.push(args.map(String).join(" "));
       });
-      const rt = await buildChatRuntime(dir, {}, SessionManager.inMemory());
+      const rt = await buildWorkspaceSessionRuntime(dir, {}, SessionManager.inMemory());
       try {
         errorSpy.mockRestore();
         expect(warnings.some((line) => line.includes('skill "greet" collision'))).toBe(true);
@@ -182,7 +182,7 @@ describe("chat: buildChatRuntime injects fastagent's assembled agent into pi's s
         `export default { description: "d", parameters: { type: "object", properties: {} }, execute: async () => ({ content: [], details: {} }) };`,
       );
 
-      const rt = await buildChatRuntime(dir, {}, SessionManager.inMemory());
+      const rt = await buildWorkspaceSessionRuntime(dir, {}, SessionManager.inMemory());
       try {
         const sp = rt.session.agent.state.systemPrompt ?? "";
         expect(sp).toContain("PERSONA_MARKER"); // ① persona from agentDir
@@ -208,7 +208,7 @@ describe("chat: buildChatRuntime injects fastagent's assembled agent into pi's s
       await writeFile(join(agentDir, "APPEND_SYSTEM.md"), "GLOBAL_APPEND_LEAK_MARKER must not reach chat.\n");
       process.env.PI_CODING_AGENT_DIR = agentDir;
 
-      const rt = await buildChatRuntime(dir, {}, SessionManager.inMemory());
+      const rt = await buildWorkspaceSessionRuntime(dir, {}, SessionManager.inMemory());
       try {
         const sp = rt.session.agent.state.systemPrompt ?? "";
         expect(sp).toContain("DEFN_ONLY_MARKER"); // fastagent's prompt is there
@@ -242,7 +242,7 @@ describe("chat: buildChatRuntime injects fastagent's assembled agent into pi's s
         `${JSON.stringify({ type: "session", version: 3, id: "other", timestamp: new Date().toISOString(), cwd: other })}\n`,
       );
 
-      const rt = await buildChatRuntime(dir, {}, SessionManager.create(dir, sessionsDir));
+      const rt = await buildWorkspaceSessionRuntime(dir, {}, SessionManager.create(dir, sessionsDir));
       try {
         let invalidated = false;
         rt.setBeforeSessionInvalidate(() => {
@@ -250,7 +250,7 @@ describe("chat: buildChatRuntime injects fastagent's assembled agent into pi's s
         });
         // A session that EXPLICITLY records another workspace is rejected before pi tears the live
         // session down — independent of process.cwd().
-        await expect(rt.importFromJsonl(imported, other)).rejects.toThrow(/fastagent chat is workspace-scoped/);
+        await expect(rt.importFromJsonl(imported, other)).rejects.toThrow(/fastagent sessions are workspace-scoped/);
         expect(invalidated).toBe(false);
         expect(rt.cwd).toBe(realpathSync(dir)); // runtime cwd is canonical (symlink-free)
       } finally {
@@ -283,7 +283,7 @@ describe("chat: buildChatRuntime injects fastagent's assembled agent into pi's s
 
       process.chdir(dir);
       const realDir = realpathSync(dir); // pi binds the realpath via process.cwd()
-      const rt = await buildChatRuntime(dir, {}, SessionManager.create(dir, sessionsDir));
+      const rt = await buildWorkspaceSessionRuntime(dir, {}, SessionManager.create(dir, sessionsDir));
       try {
         // Import the cwd-less session: no foreign cwd, so it runs in the chat workspace.
         await expect(rt.importFromJsonl(legacy)).resolves.toMatchObject({ cancelled: false });
